Expose the udns C library functions add_serv and add_serv_s to the Ruby interface.
Resolver#initialize now takes an optional hash of options; if one
or both of 'nameserver' or 'nameservers' keys are present, their
values are merged and the list of nameservers is set to the new
list (this overrides NAMESERVERS in the environment).

When passing nameservers to #initialize, "host:port" syntax is
recognized to enable querying DNS servers on non-standard ports.

Bump version to 0.3.6.2.